Solution to Check if a Number is Square of Another

+++ Check if a Number is Square of Another@@6788dd63490637cfc44e1ba4/code.c
@@ -2,9 +2,9 @@
 #include<math.h>
 int main()
 {
-    float a,b,c;
-    scanf("%f%f",&a,&b);
-    c=pow(a,0.5);
+    int a,b,c;
+    scanf("%d%d",&a,&b);
+    c=(int)pow(a,0.5);
     if(b==c){
     printf("Yes");
     }
